pyproject.toml
[tool.poetry]
name = "buenos-aires-subway-passengers-flow"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyarrow = "^15.0.2"
google-cloud-storage = "^2.16.0"
ipykernel = "^6.29.4"
pandas = "^2.2.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
